requirements.txt
# Core ML libraries
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Transformers and tokenization
transformers>=4.30.0
tokenizers>=0.13.0
sentencepiece>=0.1.99

# Data processing
datasets>=2.12.0
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# Configuration and utilities
pyyaml>=6.0
omegaconf>=2.3.0
hydra-core>=1.3.0
tqdm>=4.65.0
wandb>=0.15.0

# Development tools
pytest>=7.3.0
black>=23.3.0
isort>=5.12.0
flake8>=6.0.0
mypy>=1.3.0

# Distributed training
accelerate>=0.20.0

# Optional: for inference API
fastapi>=0.95.0
uvicorn>=0.22.0

# Jupyter for notebooks
jupyter>=1.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Memory profiling and optimization
psutil>=5.9.0
py-spy>=0.3.14